Moved data local files to backend folder
